requirements.txt
accelerate @ git+https://github.com/huggingface/accelerate.git@f20445d4acc41f3b391f523691be8c81572ed1ac
aiohttp==3.9.5
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bitsandbytes==0.42.0
bleach==6.0.0
cachetools==5.3.0
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==2.1.1
Colab-Utils @ git+https://github.com/deepvision-class/starter-code@e8d9fe711870a39796a2f8ad95538e57942d756f
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
d2l==1.0.3
datasets==2.19.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.4
fonttools==4.39.4
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.121.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud==0.34.0
google-cloud-vision==3.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
httplib2==0.22.0
huggingface-hub==0.23.0
idna==3.4
ipykernel==5.5.6
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
JPype1==1.4.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
kobert-tokenizer @ git+https://github.com/SKTBrain/KoBERT.git@47a69af87928fc24e20f571fe10c3cc9dd9af9a3#subdirectory=kobert_hf
konlpy==0.6.0
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==2.0.5
mojito2==0.1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.1
norminette==3.3.51
notebook==6.5.5
notebook_shim==0.2.2
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.2
packaging==22.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
peft==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.1.1
portpicker==1.5.2
prometheus-client==0.16.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.21.0
PyDrive==1.3.1
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.1
QtPy==2.3.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
safetensors==0.4.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
sentencepiece==0.2.0
six==1.16.0
sklearn==0.0.post5
slack_sdk==3.27.1
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
sympy==1.12
terminado==0.17.1
tinycss2==1.2.1
tokenizers==0.15.2
torch==2.0.0
torchvision==0.15.1
tornado==6.3.2
tqdm==4.66.2
traitlets==5.9.0
transformers @ git+https://github.com/huggingface/transformers.git@c8d98405a8f7b0e5d07391b671dcc61bb9d7bad5
trl==0.8.6
tweepy==3.9.0
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.13
uv==0.2.13
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==14.1
widgetsnbextension==4.0.5
xxhash==3.4.1
yarl==1.9.4
